fix(#4259): keep the guarded save out of the turn_bridge hotfile
Three gates only surfaced after the earlier one stopped failing, because
`ci-script-checks.sh` aborts at its first failing section — the same
first-failure masking that hides main's later test lanes.

1. Hot-file LOC ratchet: `turn_bridge/mod.rs` sat exactly at its 969 ceiling and
   the guard body pushed it to 995. Move the body to `inflight/save_store.rs`
   (the #4183 precedent for this exact shape) so the hotfile keeps only the call
   site: 969 -> 966. Ratchet lowered to 966 to lock the win in.
2. Test-lane coverage ratchet: `save_store::bridge_entry_guard_tests` was a new
   `cfg(test)` module no curated lane selected. Add a `test-non-pg` lane for it,
   carrying the `_pg` / `pg_` / `postgres` skips that #4953 was caused by omitting.
3. `test_fast_check_ci_wiring` pins the lane list line-for-line; add the entry.
