+++ src/services/discord/inflight.rs
@@ -154,6 +154,7 @@ pub(in crate::services::discord) use self::save_store::{
     clear_long_running_placeholder_if_matches_identity,
     mark_readopted_from_inflight_if_identity_unchanged,
     patch_restart_full_response_if_identity_unchanged, patch_restart_mode_if_matches_identity,
+    persist_bridge_entry_inflight_state,
     persist_leak_recovery_response_offset_if_matches_identity_locked,
     persist_recovery_output_path_if_matches_identity_locked,
     recovery_anchor_msg_id_if_matches_identity,
